refactor(notes): extract use_audio_import hook from NoteDetail
